working with manual text binary writing

+++ client.c
@@ -6,7 +6,7 @@
 /*   By: aantonio <aantonio@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/27 13:25:38 by aantonio          #+#    #+#             */
-/*   Updated: 2023/07/14 17:14:06 by aantonio         ###   ########.fr       */
+/*   Updated: 2023/07/15 10:42:34 by aantonio         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,10 @@ int	main(int argc, char *argv[] )
 		exit(-1);
 	}
 	server_pid = atoi(argv[1]);
-	kill(server_pid, SIGUSR1);
+	if (*argv[2] == '1')
+		kill(server_pid, SIGUSR1);
+	if (*argv[2] == '2')
+		kill(server_pid, SIGUSR2);
 	printf("%s\n",argv[2]);
 	return (0);
 }
